refactor code by creating router folder
