+++ app.js
@@ -17,6 +17,11 @@ mongoose.set('useCreateIndex', true);
 mongoose.set('useUnifiedTopology', true);
 //*************** End of correction lines to connect db correctly!!
 
+//requring routes
+var commentRoutes    = require("./routes/comments"),
+    campgroundRoutes = require("./routes/campgrounds"),
+    indexRoutes      = require("./routes/index")
+
 mongoose.connect("mongodb://localhost/natureCalling");
 app.use(bodyParser.urlencoded({extended: true}));
 app.set("view engine", "ejs");
@@ -39,144 +44,9 @@ app.use(function(req, res, next){
     next();
  });
 
-app.get("/", function(req, res) {
-    res.render("home");
-});
-
-app.get("/campgrounds", function(req, res) {
-    //call all campgrounds from DB
-    Campground.find({},function(err, allCampgrounds){
-        if(err){
-            console.log(err);
-        }
-        else{
-            res.render("campgrounds/index", {campgrounds:allCampgrounds});
-        }
-    });
-});
-
-app.get("/campgrounds/new", function(req, res){
-    res.render("campgrounds/new");
-});
-
-app.get("/campgrounds/:id",function(req, res){
-    Campground.findById(req.params.id).populate("comments").exec(function(err,foundCampgrounds){
-        if(err){
-            console.log(err);
-        }
-        else{
-            res.render("campgrounds/show", {campground:foundCampgrounds});
-        }
-    })
-});
-
-
-
-app.post("/campgrounds", function(req, res){
-    
-    var name = req.body.name;
-    var image = req.body.image;
-    var desc = req.body.description;
-    var newCampground = {name:name, image:image, description:desc};
-    
-    //Create new campground on DB
-    Campground.create(newCampground, function(err,newlyCreated){
-        if(err){
-            console.log(err);
-        }
-        else{
-            res.redirect("/campgrounds");
-        }
-    });
-});
-
-// ====================
-// COMMENTS ROUTES
-// ====================
-
-app.get("/campgrounds/:id/comments/new", isLoggedIn, function(req, res){
-    // find campground by id
-    Campground.findById(req.params.id, function(err, campground){
-        if(err){
-            console.log(err);
-        } else {
-             res.render("comments/new", {campground: campground});
-        }
-    })
-});
-
-app.post("/campgrounds/:id/comments", function(req, res){
-   //lookup campground using ID
-   Campground.findById(req.params.id, function(err, campground){
-       if(err){
-           console.log(err);
-           res.redirect("/campgrounds");
-       } else {
-        Comment.create(req.body.comment, function(err, comment){
-           if(err){
-               console.log(err);
-           } else {
-               campground.comments.push(comment);
-               campground.save();
-               res.redirect('/campgrounds/' + campground._id);
-           }
-        });
-       }
-   });
-   //create new comment
-   //connect new comment to campground
-   //redirect campground show page
-});
-//  ===========
-// AUTH ROUTES
-//  ===========
-
-// show register form
-app.get("/register", function(req, res){
-    res.render("register"); 
- });
- //handle sign up logic
- app.post("/register", function(req, res){
-     var newUser = new User({username: req.body.username});
-     User.register(newUser, req.body.password, function(err, user){
-         if(err){
-             console.log(err);
-             return res.render("register");
-         }
-         passport.authenticate("local")(req, res, function(){
-            res.redirect("/campgrounds"); 
-         });
-     });
- });
- 
-// show login form
-app.get("/login", function(req, res){
-    res.render("login"); 
- });
- // handling login logic
- app.post("/login", passport.authenticate("local", 
-     {
-         successRedirect: "/campgrounds",
-         failureRedirect: "/login"
-     }), function(req, res){
- });
- 
- // logic route
- app.get("/logout", function(req, res){
-    req.logout();
-    res.redirect("/campgrounds");
- });
- 
- function isLoggedIn(req, res, next){
-     if(req.isAuthenticated()){
-         return next();
-     }
-     res.redirect("/login");
- }
-
-// app.get("*",function(req,res){
-//     res.send("Send something else.");
-// });
+app.use("/", indexRoutes);
+app.use("/campgrounds", campgroundRoutes);
+app.use("/campgrounds/:id/comments", commentRoutes);
 
 var port = 3000;
 app.listen(port, function() {
